+++ ModCore/Commands/Owner.cs
@@ -1,9 +1,9 @@
-﻿using DSharpPlus.CommandsNext;
+﻿using System;
+using System.Threading.Tasks;
+using DSharpPlus.CommandsNext;
 using DSharpPlus.CommandsNext.Attributes;
 using DSharpPlus.Interactivity;
 using ModCore.Entities;
-using System;
-using System.Threading.Tasks;
 
 namespace ModCore.Commands
 {
